fix ui not displaying properly in bigger screens
