+++ client/src/pages/Budgets/styles.ts
@@ -1,4 +1,4 @@
-import { styled, Stack, Box } from '@mui/system';
+import { styled, Stack } from '@mui/system';
 
 export const BudgetStack = styled(Stack)(({ theme }) => ({
   background: `${theme.palette.secondary.main} 0%`,
@@ -13,13 +13,3 @@ export const BudgetStack = styled(Stack)(({ theme }) => ({
   paddingTop: '2vh',
   paddingBottom: '2vh',
 }));
-
-export const PopupDiv = styled(Box)({
-  maxHeight: '100vh',
-  display: 'flex',
-  margin: '1px',
-  paddingTop: '2vh',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
