+++ client/src/pages/Home/styles.ts
@@ -2,10 +2,12 @@ import { styled, Stack } from '@mui/system';
 
 export const HomePageStack = styled(Stack)({
   display: 'flex',
-  paddingLeft: '5%',
-  paddingRight: '5%',
-  paddingBottom: '5%',
   flexWrap: 'wrap',
-  justifyContent: 'center',
+  flexDirection: 'row',
   alignItems: 'center',
+  justifyContent: 'center',
+  margin: '0rem',
+  minWidth: '800px',
+  maxWidth: '1000px',
+  width: '80vw',
 });
