+++ client/src/pages/UserSettings/styles.ts
@@ -14,13 +14,3 @@ export const SettingsStack = styled(Stack)(({ theme }) => ({
   paddingTop: '2vh',
   paddingBottom: '2vh',
 }));
-
-export const PopupDiv = styled(Box)({
-  maxHeight: '100vh',
-  display: 'flex',
-  margin: '1px',
-  paddingTop: '10vh',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
