+++ client/src/styles.ts
@@ -1,0 +1,11 @@
+import { styled, Stack, Box } from '@mui/system';
+
+export const PopupDiv = styled(Box)({
+  maxHeight: '100vh',
+  display: 'flex',
+  margin: '1px',
+  paddingTop: '2vh',
+  flexDirection: 'column',
+  justifyContent: 'center',
+  alignItems: 'center',
+});
